rename the shared folder with modal_shared
