+++ src/components/redirect_notification/MobileNotification.js
@@ -1,9 +1,9 @@
 import * as s from "./RedirectionNotificationModal.sc";
 import { useState } from "react";
-import Modal from "./shared/Modal";
-import Header from "./shared/Header";
-import Body from "./shared/Body";
-import Footer from "./shared/Footer";
+import Modal from "./modal_shared/Modal";
+import Header from "./modal_shared/Header";
+import Body from "./modal_shared/Body";
+import Footer from "./modal_shared/Footer";
 import BookmarkBorderIcon from "@mui/icons-material/BookmarkBorder";
 
 export default function MobileNotification({
