+++ src/components/redirect_notification/modal_shared/Body.js
