+++ src/components/redirect_notification/modal_shared/Footer.js
