+++ src/components/redirect_notification/modal_shared/Header.js
