+++ src/components/redirect_notification/modal_shared/Modal.js
